+++ examples/esbuild/src/dyn.ts
@@ -1,0 +1,3 @@
+const myString = "This is a dynamic import!";
+
+export default myString;