+++ examples/esbuild/src/mod.ts
@@ -4,7 +4,7 @@ import { basename } from "https://deno.land/std@0.100.0/path/mod.ts";
 export async function main(): Promise<void> {
   console.log(myString);
   console.log((await import("./dyn.ts")).default);
-  console.log(`Current directory name: ${basename(Deno.cwd())}.`);
+  console.log(`Current directory: ${basename(Deno.cwd())}.`);
 }
 
 export default main;