+++ plugins/importmap/test/fixtures/modules/simple/app/views.js
@@ -1,4 +1,4 @@
-import { css, html } from "lit-element";
+import { html } from "lit-element";
 
 export default function view(items) {
   return html`<p>Hello ${items[0]}!</p>`;